+++ src/components/button/index.ts
@@ -1,0 +1,6 @@
+export * from './basic'
+export * from './icon'
+export * from './link'
+export * from './trade'
+export * from './social'
+export * from './types'