+++ src/components/input/index.ts
@@ -1,1 +1,2 @@
 export * from './search-field'
+export * from './text-field'